feat: implements multi var in the key for the keyval command
The idea is very basic. We make a string mask (intermediate string).
Where there's a variable, put a $ and take the variable name to get its index.
Other chars are just copied.
During running time, all the '$' in string mask are replaced by the value of the desirable variable.
The variable that store the index for the only variable we could have before
has been changed for a array with size 15, which is the max variables allowed in the keyval key-string.