feat: implements timeout for keyval
Using the timer for events implemented by nginx, we can use ttl
in keyval zones. This creates a struct for store the pointer for the node, which
should be deleted after an amount of time specified on the configuration, and
the pointer for the context. The ttl belongs to this context too.
When the node is created, the function ngx_keyval_shm_set_data creates an event
and adds a timer for this event. The handler (ngx_keyval_delete_timeout_node_shm)
deletes the node. So, the next time when key is checked, there won't be a node with it.